Upgrade dropwizard and other dependencies.
Given that active development on coner is about to startup again, I went
ahead and updated a lot of other dependencies.  I also made a lot of
checkstyle rules 'error' since we were already compliant and only marked
the few that we don't care about as 'warning'.

There were some changes necessary to get the DAO unit tests to work but
when the new DAOTestRule:

https://github.com/dropwizard/dropwizard/blob/master/dropwizard-testing/src/main/java/io/dropwizard/testing/junit/DAOTestRule.java

is integrated into a release of dropwizard-testing, we can make use of
it in a very clean way.

Resolves #126